Fixed amount of fireball damage done to allies, close issue #356
In the original game the whole series did damage 0.02 (0.10 * 0.2) to human. In Gold, one particle do such damage, so 51 particles do 51 * 0.02 = 1.02 damage to human! I changed the multiplier for allies from 0.2 to 0.004 (approx. 0.2 / 51), but this problem probably occured somewhere else during rewriting the engine from DirectX to OpenGL and SDL.