Migrate to GitHub Actions
This is pretty much a straight port of our Jenkinsfile to GitHub
Actions. It could use some modernization in a few places, but it should
work for now.

Note that I left the Jenkinsfile for now - we'll be using them both
until we are sure that the GitHub Actions version is stable.