Simplyfy FlatFreeSpace
The "from external to internal" branch never worked. Given that
it was broken ever since the initial introduction of flatspace()
and no one noticed or complained I thik it can simply be removed.
The only case where the "from external to internal" branch would find
anything is if `minRadius == maxRadius` otherwise the search would terminate
immediately. The code can already handle `minRadius == maxRadius` so
deleting the branch should lead to no loss of functionality.

How was it broken?
```
if (minRadius < maxRadius) {
   // ...
} else {
   // Here `minRadius >= maxRadius`
   // therefore `radius <= maxRadius` is only true if `minRadius == maxRadius`
   // otherwise the search terminates immediately.
   for (float radius = minRadius; radius <= maxRadius; radius += space) {
      // ...
   }
}
```