Save level-supplied programs when saving the game
CBot scripts that are loaded from the game level or from a mod
(such as the scripts implementing enemy AI) may change from one version
of the game or mod to the next.

Instead of preserving those scripts as part of the save file we were
only storing file names and loading new modified versions of those
scripts when when loading a save.

When saving the game we save the current execution state of all running
programs. Attempting to continue running the program from where we
left off after the source code changed may cause the script to behave
strangely or even cause **the whole game** to crash.

I fixed this by saving all scripts when saving the game. This way the
saved execution state always matches the source code.

Fixes https://github.com/colobot/colobot/issues/1623
Undoes "Never save copy of programs from level files when saving scene" 4182be321695306385543c0584cb780959ceb25e

Legacy save files do not contain some of the source code that was
available when the game was saved. The version of the mod/game may have
changed since - we don't know. Those files may (in general) contain
classes and public functions. To make sure that loading a legacy save
file does not crash the game we discard the current execution state of
all scripts and start them over from the beginning.

Unlike player-written programs, level-supplied programs are not saved to
be loaded when the level is restarted. This should prevent
re-introduction of https://github.com/colobot/colobot/issues/534